+++ edx/analytics/tasks/common/spark.py
@@ -1,8 +1,6 @@
 from edx.analytics.tasks.common.pathutil import EventLogSelectionDownstreamMixin, PathSelectionByDateIntervalTask
 from edx.analytics.tasks.util.overwrite import OverwriteOutputMixin
-import edx.analytics.tasks.util.opaque_key_util as opaque_key_util
 from luigi.contrib.spark import PySparkTask
-from edx.analytics.tasks.util.constants import PredicateLabels
 
 
 class EventLogSelectionMixinSpark(EventLogSelectionDownstreamMixin):
@@ -134,23 +132,27 @@ class SparkJobTask(OverwriteOutputMixin, PySparkTask):
             zipfile = ZipFile(zipfile_path, 'w', ZIP_DEFLATED)
             zipdir(mod_path, zipfile, package)
             zipfile.close()
-            print '\n LOADING via spark context {}\n'.format(zipfile_path)
             self._spark_context.addPyFile(zipfile_path)
 
 
     def _load_internal_dependency_on_cluster(self):
-        """creates a zip of edx dir and loads it on spark worker nodes"""
+        """creates a zip of edx and luigi package and loads it on spark worker nodes"""
         # TODO: delete zipfile after loading on cluster completes
         import os
         import tempfile
         import shutil
         import edx as import_dir_path
-        package_name = 'edx_analytics_tasks'
+        import luigi
         tmp_dir = tempfile.mkdtemp()
-        archive_dir = os.path.join(import_dir_path.__path__[0], '../')  # more reasonable than using path with __file__
-        zipfile_path = shutil.make_archive(os.path.join(tmp_dir, package_name), 'zip', archive_dir, 'edx/')
+        # zip edx package
+        archive_dir = os.path.join(import_dir_path.__path__[0], '../')
+        zipfile_edx = shutil.make_archive(os.path.join(tmp_dir, 'edx_analytics_tasks'), 'zip', archive_dir, 'edx/')
+        # zip luigi package
+        archive_dir = os.path.join(luigi.__path__[0], '../')
+        zipfile_luigi = shutil.make_archive(os.path.join(tmp_dir, 'luigi'), 'zip', archive_dir, 'luigi/')
         # add zipfile to spark context
-        self._spark_context.addPyFile(zipfile_path)
+        self._spark_context.addPyFile(zipfile_edx)
+        self._spark_context.addPyFile(zipfile_luigi)
 
     def run(self):
         self.remove_output_on_overwrite()
@@ -158,79 +160,6 @@ class SparkJobTask(OverwriteOutputMixin, PySparkTask):
 
     def main(self, sc, *args):
         self.init_spark(sc)
-        # self._load_internal_dependency_on_cluster()  # load internal dependency for spark worker nodes on cluster
-        self._load_external_dependency_on_cluster()  # load external dependency for spark worker nodes on cluster
+        self._load_internal_dependency_on_cluster()  # load internal dependency for spark worker nodes on cluster
+        # self._load_external_dependency_on_cluster()  # load external dependency for spark worker nodes on cluster
         self.spark_job()
-
-
-def get_event_predicate_labels(event):
-    """Creates labels by applying hardcoded predicates to a single event."""
-    # We only want the explicit event, not the implicit form.
-    # return 'test'
-    event_type = event['event_type']
-    event_source = event['event_source']
-
-    labels = PredicateLabels.ACTIVE_LABEL
-
-    # task & enrollment events are filtered out by spark later as it speeds up due to less # of records
-
-    if event_source == 'server':
-        if event_type == 'problem_check':
-            labels += ',' + PredicateLabels.PROBLEM_LABEL
-
-        if event_type.startswith('edx.forum.') and event_type.endswith('.created'):
-            labels += ',' + PredicateLabels.POST_FORUM_LABEL
-
-    if event_source in ('browser', 'mobile'):
-        if event_type == 'play_video':
-            labels += ',' + PredicateLabels.PLAY_VIDEO_LABEL
-
-    return labels
-
-
-def get_key_value_from_event(event, key, default_value=None):
-    """
-    Get value from event dict by key
-    Pyspark does not support dict.get() method, so this approach seems reasonable
-    """
-    try:
-        default_value = event[key]
-    except KeyError:
-        pass
-    return default_value
-
-
-def get_course_id(event, from_url=False):
-    """Gets course_id from event's data."""
-
-    # Get the event data:
-    event_context = get_key_value_from_event(event, 'context')
-    if event_context is None:
-        # Assume it's old, and not worth logging...
-        return ''
-
-    # Get the course_id from the data, and validate.
-    course_id = opaque_key_util.normalize_course_id(get_key_value_from_event(event_context, 'course_id', ''))
-    if course_id:
-        if opaque_key_util.is_valid_course_id(course_id):
-            return course_id
-        else:
-            return ''  # we'll filter out empty course since string is expected
-
-    # Try to get the course_id from the URLs in `event_type` (for implicit
-    # server events) and `page` (for browser events).
-    if from_url:
-        source = get_key_value_from_event(event, 'event_source')
-
-        if source == 'server':
-            url = get_key_value_from_event(event, 'event_type', '')
-        elif source == 'browser':
-            url = get_key_value_from_event(event, 'page', '')
-        else:
-            url = ''
-
-        course_key = opaque_key_util.get_course_key_from_url(url)
-        if course_key:
-            return unicode(course_key)
-
-    return ''
